+++ src/apps/hector-network/fantom/hector-network.bond.contract-position-fetcher.ts
@@ -4,12 +4,12 @@ import { APP_TOOLKIT, IAppToolkit } from '~app-toolkit/app-toolkit.interface';
 import { Register } from '~app-toolkit/decorators';
 import { getImagesFromToken, getLabelFromToken } from '~app-toolkit/helpers/presentation/image.present';
 import { MetaType } from '~position/position.interface';
+import { ContractPositionTemplatePositionFetcher } from '~position/template/contract-position.template.position-fetcher';
 import {
-  ContractPositionTemplatePositionFetcher,
-  DisplayPropsStageParams,
-  GetTokenBalancesPerPositionParams,
-  TokenStageParams,
-} from '~position/template/contract-position.template.position-fetcher';
+  GetDisplayPropsParams,
+  GetTokenBalancesParams,
+  GetTokenDefinitionsParams,
+} from '~position/template/contract-position.template.types';
 import { Network } from '~types';
 
 import { HectorNetworkBondDepository, HectorNetworkContractFactory } from '../contracts';
@@ -21,9 +21,9 @@ const network = Network.FANTOM_OPERA_MAINNET;
 
 @Register.ContractPositionFetcher({ appId, groupId, network })
 export class FantomHectorNetworkBondContractPositionFetcher extends ContractPositionTemplatePositionFetcher<HectorNetworkBondDepository> {
-  appId = HECTOR_NETWORK_DEFINITION.id;
-  groupId = HECTOR_NETWORK_DEFINITION.groups.bond.id;
-  network = Network.FANTOM_OPERA_MAINNET;
+  appId = appId;
+  groupId = groupId;
+  network = network;
   groupLabel = 'Bonds';
 
   constructor(
@@ -33,7 +33,7 @@ export class FantomHectorNetworkBondContractPositionFetcher extends ContractPosi
     super(appToolkit);
   }
 
-  async getDescriptors() {
+  async getDefinitions() {
     return [
       { address: '0x4099eb0e82ffa0048e4bf037a9743ca05ec561d7' },
       { address: '0x6c9b3a47a28a39fea65e99d97895e717df1706d0' },
@@ -43,6 +43,9 @@ export class FantomHectorNetworkBondContractPositionFetcher extends ContractPosi
       { address: '0xde02631d898acd1bb8ff928c0f0ffa0cf29ab374' },
       { address: '0xa695750b8439ab2afbd88310946c99747c5b3a2e' },
       { address: '0x72de9f0e51ca520379a341318870836fdcaf03b9' },
+      { address: '0xdd62c045d9a873f1206a5291dcf0ea9fc2aa8ddf' },
+      { address: '0x4441f551001ab0785f1006929aa86d0c846f30cc' },
+      { address: '0x312ade5a805e5f3975bbdbb9feb5ef4d1e15eb8f' },
     ];
   }
 
@@ -50,7 +53,7 @@ export class FantomHectorNetworkBondContractPositionFetcher extends ContractPosi
     return this.contractFactory.hectorNetworkBondDepository({ address, network: this.network });
   }
 
-  async getTokenDescriptors({ contract }: TokenStageParams<HectorNetworkBondDepository>) {
+  async getTokenDefinitions({ contract }: GetTokenDefinitionsParams<HectorNetworkBondDepository>) {
     const [principle, claimable] = await Promise.all([contract.principle(), contract.HEC()]);
 
     return [
@@ -60,11 +63,11 @@ export class FantomHectorNetworkBondContractPositionFetcher extends ContractPosi
     ];
   }
 
-  async getLabel({ contractPosition }: DisplayPropsStageParams<HectorNetworkBondDepository>) {
+  async getLabel({ contractPosition }: GetDisplayPropsParams<HectorNetworkBondDepository>) {
     return `${getLabelFromToken(contractPosition.tokens[2])} Bond`;
   }
 
-  async getImages({ contractPosition }: DisplayPropsStageParams<HectorNetworkBondDepository>) {
+  async getImages({ contractPosition }: GetDisplayPropsParams<HectorNetworkBondDepository>) {
     return getImagesFromToken(contractPosition.tokens[2]);
   }
 
@@ -72,7 +75,7 @@ export class FantomHectorNetworkBondContractPositionFetcher extends ContractPosi
     address,
     contract,
     multicall,
-  }: GetTokenBalancesPerPositionParams<HectorNetworkBondDepository>) {
+  }: GetTokenBalancesParams<HectorNetworkBondDepository>) {
     const [bondInfo, claimablePayout] = await Promise.all([
       multicall.wrap(contract).bondInfo(address),
       multicall.wrap(contract).pendingPayoutFor(address),