+++ src/apps/hector-network/hector-network.definition.ts
@@ -28,6 +28,12 @@ export const HECTOR_NETWORK_DEFINITION = appDefinition({
       label: 'Wrapped sHEC V2',
     },
 
+    hec: {
+      id: 'hec',
+      type: GroupType.TOKEN,
+      label: 'Hector Network',
+    },
+
     bond: {
       id: 'bond',
       type: GroupType.POSITION,
@@ -39,6 +45,12 @@ export const HECTOR_NETWORK_DEFINITION = appDefinition({
       type: GroupType.POSITION,
       label: 'Bonds',
     },
+
+    farm: {
+      id: 'farm',
+      type: GroupType.POSITION,
+      label: 'Farms',
+    },
   },
   links: {
     twitter: 'https://twitter.com/Hector_Network',
@@ -49,6 +61,7 @@ export const HECTOR_NETWORK_DEFINITION = appDefinition({
   },
   supportedNetworks: {
     [Network.FANTOM_OPERA_MAINNET]: [AppAction.VIEW],
+    [Network.BINANCE_SMART_CHAIN_MAINNET]: [AppAction.VIEW],
   },
 });