+++ src/position/template/single-staking.dynamic.template.contract-position-fetcher.ts
@@ -6,14 +6,14 @@ import { APP_TOOLKIT, IAppToolkit } from '~app-toolkit/app-toolkit.interface';
 import { getImagesFromToken, getLabelFromToken } from '~app-toolkit/helpers/presentation/image.present';
 import { MetaType } from '~position/position.interface';
 import { isClaimable, isSupplied } from '~position/position.utils';
+import { ContractPositionTemplatePositionFetcher } from '~position/template/contract-position.template.position-fetcher';
 import {
-  ContractPositionTemplatePositionFetcher,
-  DataPropsStageParams,
-  DisplayPropsStageParams,
-  GetTokenBalancesPerPositionParams,
-  TokenStageParams,
-  UnderlyingTokenDescriptor,
-} from '~position/template/contract-position.template.position-fetcher';
+  GetDataPropsParams,
+  GetDisplayPropsParams,
+  GetTokenBalancesParams,
+  GetTokenDefinitionsParams,
+  UnderlyingTokenDefinition,
+} from '~position/template/contract-position.template.types';
 
 export type SingleStakingFarmDataProps = {
   liquidity: number;
@@ -30,32 +30,30 @@ export abstract class SingleStakingFarmDynamicTemplateContractPositionFetcher<
   }
 
   abstract getFarmAddresses(): string[] | Promise<string[]>;
-  abstract getStakedTokenAddress(params: TokenStageParams<T, V>): Promise<string>;
-  abstract getRewardTokenAddresses(params: TokenStageParams<T, V>): Promise<string | string[]>;
-  abstract getRewardRates(params: DataPropsStageParams<T, V>): Promise<BigNumberish | BigNumberish[]>;
-  abstract getStakedTokenBalance(
-    params: GetTokenBalancesPerPositionParams<T, SingleStakingFarmDataProps>,
-  ): Promise<BigNumberish>;
+  abstract getStakedTokenAddress(params: GetTokenDefinitionsParams<T>): Promise<string>;
+  abstract getRewardTokenAddresses(params: GetTokenDefinitionsParams<T>): Promise<string | string[]>;
+  abstract getRewardRates(params: GetDataPropsParams<T, V>): Promise<BigNumberish | BigNumberish[]>;
+  abstract getStakedTokenBalance(params: GetTokenBalancesParams<T, SingleStakingFarmDataProps>): Promise<BigNumberish>;
   abstract getRewardTokenBalances(
-    params: GetTokenBalancesPerPositionParams<T, SingleStakingFarmDataProps>,
+    params: GetTokenBalancesParams<T, SingleStakingFarmDataProps>,
   ): Promise<BigNumberish | BigNumberish[]>;
 
-  async getDescriptors() {
+  async getDefinitions() {
     const farmAddresses = await this.getFarmAddresses();
     return farmAddresses.map(address => ({ address }));
   }
 
-  async getTokenDescriptors(params: TokenStageParams<T, V>) {
+  async getTokenDefinitions(params: GetTokenDefinitionsParams<T>) {
     const stakedTokenAddress = await this.getStakedTokenAddress(params);
     const rewardTokenAddresses = await this.getRewardTokenAddresses(params).then(v => (isArray(v) ? v : [v]));
 
-    const tokens: UnderlyingTokenDescriptor[] = [];
+    const tokens: UnderlyingTokenDefinition[] = [];
     tokens.push({ metaType: MetaType.SUPPLIED, address: stakedTokenAddress.toLowerCase() });
     tokens.push(...rewardTokenAddresses.map(v => ({ metaType: MetaType.CLAIMABLE, address: v.toLowerCase() })));
     return tokens;
   }
 
-  async getDataProps(params: DataPropsStageParams<T, V>): Promise<V> {
+  async getDataProps(params: GetDataPropsParams<T, V>): Promise<V> {
     const { contractPosition, multicall } = params;
     const stakedToken = contractPosition.tokens.find(isSupplied)!;
     const rewardTokens = contractPosition.tokens.filter(isClaimable);
@@ -76,16 +74,16 @@ export abstract class SingleStakingFarmDynamicTemplateContractPositionFetcher<
     return { liquidity, apy, isActive } as V;
   }
 
-  async getLabel({ contractPosition }: DisplayPropsStageParams<T>) {
+  async getLabel({ contractPosition }: GetDisplayPropsParams<T>) {
     const stakedToken = contractPosition.tokens.find(isSupplied)!;
     return getLabelFromToken(stakedToken);
   }
 
-  async getImages({ contractPosition }: DisplayPropsStageParams<T, V>) {
+  async getImages({ contractPosition }: GetDisplayPropsParams<T, V>) {
     return contractPosition.tokens.filter(isSupplied).flatMap(v => getImagesFromToken(v));
   }
 
-  async getTokenBalancesPerPosition(params: GetTokenBalancesPerPositionParams<T, SingleStakingFarmDataProps>) {
+  async getTokenBalancesPerPosition(params: GetTokenBalancesParams<T, SingleStakingFarmDataProps>) {
     const tokenBalances: BigNumberish[] = [];
 
     const [stakedBalanceRaw, rewardTokenBalancesRaw] = await Promise.all([